refactor(polls.PollResults): Migrate PollResults logic to an abstract component
